Внести правки.

+++ src/main/java/ru/practicum/shareit/booking/service/BookingServiceImpl.java
@@ -31,9 +31,12 @@ public class BookingServiceImpl implements BookingService {
     @Override
     public BookingDto createBooking(RequestBookingDto bookingDto, Long userId) {
         Item item = itemDao.findById(bookingDto.getItemId())
-                .orElseThrow(() -> new NotFoundException("Вещь не найден."));
+                .orElseThrow(() -> new NotFoundException("Вещь c id " + bookingDto.getItemId() + " не найден."));
         if (item.isAvailable()) {
-            User user = userDao.findById(userId).orElseThrow(() -> new NotFoundException("Пользователь не найден."));
+            User user = findUserById(userId);
+            if (bookingDto.getStart().isAfter(bookingDto.getEnd())) {
+                throw new ValidationException("Дата начала бронирования не может быть позже, конца бронирования");
+            }
             log.info("Передаём запрос на создание бронирования в bookingDao.");
             return BookingMapper.toBookingDto(bookingDao.save(BookingMapper.fromBookingDto(bookingDto, user, item)));
         }
@@ -43,8 +46,7 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public BookingDto updateBookingStatus(Long bookingId, boolean approveState, Long userId) {
-        Booking oldBooking = bookingDao.findById(bookingId)
-                .orElseThrow(() -> new NotFoundException("Бронь не найден."));
+        Booking oldBooking = findBookingById(bookingId);
         if (oldBooking.getItem().getOwner().getId().equals(userId)) {
             if (approveState) {
                 oldBooking.setStatus(Status.APPROVED);
@@ -62,8 +64,8 @@ public class BookingServiceImpl implements BookingService {
     @Override
     public BookingDto getBooking(Long bookingId, Long userId) {
         log.info("Передаём запрос на получение информации о бронирование с id{} в bookingDao.", bookingId);
-        Booking booking = bookingDao.findById(bookingId)
-                .orElseThrow(() -> new NotFoundException("Бронь не найден."));
+        findUserById(userId);
+        Booking booking = findBookingById(bookingId);
         if (booking.getItem().getOwner().getId().equals(userId) || booking.getBooker().getId().equals(userId)) {
             return BookingMapper.toBookingDto(booking);
         }
@@ -73,7 +75,7 @@ public class BookingServiceImpl implements BookingService {
 
     @Override
     public Collection<BookingDto> getUserBookings(String state, Long userId) {
-        User user = userDao.findById(userId).orElseThrow(() -> new NotFoundException("Пользователь не найден."));
+        findUserById(userId);
         log.info("Передаём запрос на получение информации о бронирование пользователя с id{} в bookingDao.", userId);
         Collection<Booking> requiredBookings = switch (state) {
             case "ALL" -> bookingDao.findAllUserBookingByUserId(userId);
@@ -84,12 +86,14 @@ public class BookingServiceImpl implements BookingService {
             case "REJECTED" -> bookingDao.findRejectedUserBookingByUserId(userId);
             default -> throw new ValidationException("Неверный параметр состояния: " + state);
         };
-        return requiredBookings.stream().map(BookingMapper::toBookingDto).toList();
+        return requiredBookings.stream()
+                .map(BookingMapper::toBookingDto)
+                .toList();
     }
 
     @Override
     public Collection<BookingDto> getOwnerBookings(String state, Long userId) {
-        User user = userDao.findById(userId).orElseThrow(() -> new NotFoundException("Пользователь не найден."));
+        findUserById(userId);
         Collection<Long> userItems = itemDao.findByOwnerId(userId).stream()
                 .map(Item::getId)
                 .toList();
@@ -106,12 +110,29 @@ public class BookingServiceImpl implements BookingService {
                 default -> throw new ValidationException("Неверный параметр состояния: " + state);
             };
         }
-        return requiredBookings.stream().map(BookingMapper::toBookingDto).toList();
+        return requiredBookings.stream()
+                .map(BookingMapper::toBookingDto)
+                .toList();
     }
 
     @Override
-    public void deleteBooking(Long bookingId) {
+    public void deleteBooking(Long bookingId, Long userId) {
         log.info("Передаём запрос на удаление бронирования с id{} в bookingDao.", bookingId);
+        Booking booking = findBookingById(bookingId);
+        findUserById(userId);
+        if (!booking.getBooker().getId().equals(userId)) {
+            throw new ValidationException("Только создателя бронирования может его удалить");
+        }
         bookingDao.deleteById(bookingId);
+    }
+
+    private User findUserById(Long userId) {
+        return userDao.findById(userId)
+                .orElseThrow(() -> new NotFoundException("Пользователь c id " + userId + " не найден."));
+    }
+
+    private Booking findBookingById(Long bookingId) {
+        return bookingDao.findById(bookingId)
+                .orElseThrow(() -> new NotFoundException("Бронь c id " + bookingId + " не найдена."));
     }
 }
